LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioparam-interface/k-rate-dynamics-compressor-connections-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
FAIL Executing "attack" promise_test: Unhandled rejection with value: object "TypeError: function is not a constructor (evaluating 'new ChannelMergerNode(
            context, {numberOfInputs: context.destination.channelCount})')"
FAIL Executing "knee" promise_test: Unhandled rejection with value: object "TypeError: function is not a constructor (evaluating 'new ChannelMergerNode(
            context, {numberOfInputs: context.destination.channelCount})')"
FAIL Executing "ratio" promise_test: Unhandled rejection with value: object "TypeError: function is not a constructor (evaluating 'new ChannelMergerNode(
            context, {numberOfInputs: context.destination.channelCount})')"
FAIL Executing "release" promise_test: Unhandled rejection with value: object "TypeError: function is not a constructor (evaluating 'new ChannelMergerNode(
            context, {numberOfInputs: context.destination.channelCount})')"
FAIL Executing "threshold" promise_test: Unhandled rejection with value: object "TypeError: function is not a constructor (evaluating 'new ChannelMergerNode(
            context, {numberOfInputs: context.destination.channelCount})')"
PASS Audit report 
PASS > [attack] Dynamics compressor attack 
PASS > [knee] Dynamics compressor knee 
PASS > [ratio] Dynamics compressor ratio 
PASS > [release] Dynamics compressor release 
PASS > [threshold] Dynamics compressor threshold 
PASS # AUDIT TASK RUNNER FINISHED: 5 tasks ran successfully. 
